# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Alexandryn system API
  version: 0.0.0
  description: >-
    The contract between the Go server and every client (the Electron
    renderer and any LAN browser). Errors share one shape across every
    endpoint.
servers:
  - url: /
    description: The Go server's own origin (loopback by default, constitution §6).

# Path-based versioning. Health lives at the root, outside the versioned
# surface, because it predates and is independent of it.
paths:
...
  /api/v1/auth/setup:
    post:
      operationId: setupAdmin
      summary: Initialize system with master admin account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '201':
          description: Master admin registered successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              example:
                user:
                  id: "00000000-0000-0000-0000-000000000001"
                  username: "librarian"
                  email: "admin@alexandryn.org"
                  role: "admin"
                accessToken: "mock.jwt.token"
                refreshToken: "mock-refresh-token"
        '409':
          description: System is already initialized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: conflict
                message: "system is already initialized"
                correlationId: "00000000-0000-0000-0000-000000000000"
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate user with credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailOrUsername, password]
              properties:
                emailOrUsername: { type: string }
                password: { type: string }
                enrolmentGrant: { type: string }
      responses:
        '200':
          description: Authentication successful or MFA required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              example:
                user:
                  id: "00000000-0000-0000-0000-000000000001"
                  username: "librarian"
                  email: "admin@alexandryn.org"
                  role: "admin"
                accessToken: "mock.jwt.token"
                refreshToken: "mock-refresh-token"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: unauthorized
                message: "invalid username or password"
                correlationId: "00000000-0000-0000-0000-000000000000"
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate refresh token and issue new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Fresh access token and rotated refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              example:
                accessToken: "mock.jwt.token2"
                refreshToken: "mock-refresh-token2"
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: unauthorized
                message: "invalid refresh token"
                correlationId: "00000000-0000-0000-0000-000000000000"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke refresh token and end session
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '204':
          description: Logged out successfully
  /api/v1/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      summary: Request a password reset link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Password reset link dispatched if account exists
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
              example:
                message: "if the email is registered, a password reset link has been dispatched"
  /api/v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Confirm password reset and set new password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200':
          description: Password reset completed successfully
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
              example:
                success: true
...
  /api/v1/auth/mfa/totp/verify:
    post:
      operationId: verifyTOTP
      summary: Verify 2FA challenge code during login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfaTicket]
              properties:
                mfaTicket: { type: string }
                code: { type: string }
                recoveryCode: { type: string }
                enrolmentGrant:
                  type: string
                  description: >-
                    Resent from the login step when MFA is enabled, so device
                    pairing completes once login actually finishes rather than
                    on the password-only step.
      responses:
        '200':
          description: 2FA verification successful, tokens issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
              example:
                user:
                  id: "00000000-0000-0000-0000-000000000001"
                  username: "librarian"
                  email: "admin@alexandryn.org"
                  role: "admin"
                accessToken: "mock.jwt.token"
                refreshToken: "mock-refresh-token"
...
  /api/v1/libraries:
    get:
      operationId: listLibraries
      summary: List libraries accessible to current user
      responses:
        '200':
          description: Accessible libraries
          content:
            application/json:
              schema:
                type: object
                required: [libraries]
                properties:
                  libraries:
                    type: array
                    items: { $ref: '#/components/schemas/Library' }
              example:
                libraries:
                  - id: "00000000-0000-0000-0000-000000000001"
                    name: "Default Library"
                    description: "Main library namespace"
                    allowReaderUploads: false
                    createdAt: "2026-09-01T12:00:00Z"
                    updatedAt: "2026-09-01T12:00:00Z"
    post:
      operationId: createLibrary
      summary: Create a new library namespace (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                allowReaderUploads: { type: boolean }
      responses:
        '201':
          description: Library created successfully
          content:
            application/json:
              schema:
                type: object
                required: [library]
                properties:
                  library: { $ref: '#/components/schemas/Library' }
              example:
                library:
                  id: "00000000-0000-0000-0000-000000000002"
                  name: "Comics"
                  description: "Comics and Manga"
                  allowReaderUploads: true
                  createdAt: "2026-09-01T12:00:00Z"
                  updatedAt: "2026-09-01T12:00:00Z"
...
  /api/v1/network/pair/verify:
    post:
      operationId: verifyPairing
      summary: Verify candidate device pairing code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                label: { type: string }
      responses:
        '200':
          description: Pairing code verified and enrolment grant issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyPairingResponse' }
              example:
                enrolmentGrant: "mock.enrolment.jwt"
                address: "192.168.1.50:8080"
                hostName: "alexandryn.local"
        '400':
          description: Malformed code format
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: invalid_input
                message: "malformed pairing code: must be 8 Crockford base32 characters"
                correlationId: "00000000-0000-0000-0000-000000000000"
        '403':
          description: Origin not allowed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: forbidden
                message: request origin is not allowed
                correlationId: "00000000-0000-0000-0000-000000000000"
        '404':
          description: Pairing code not recognized or expired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: not_found
                message: pairing code not recognised
                correlationId: "00000000-0000-0000-0000-000000000000"
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: unsupported_media_type
                message: Content-Type must be application/json
                correlationId: "00000000-0000-0000-0000-000000000000"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: rate_limited
                message: too many requests, slow down and try again shortly
                correlationId: "00000000-0000-0000-0000-000000000000"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId]
      description: >-
        The one error shape every endpoint uses. `correlationId` matches
        the id in the server's structured logs for the same request.
      properties:
        code:
          type: string
          description: Machine-readable category (domain.Category).
          example: not_found
        message:
          type: string
          description: >-
            Human-readable, specific, no apology, says what to do next
            (constitution §11).
          example: no such endpoint
        correlationId:
          type: string
          description: Ties this response to one server log line.
          example: 00000000-0000-0000-0000-000000000000

    # ── Library domain schemas ───────────────────────────────────────
...
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, username, email, role]
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, reader] }
    AuthResponse:
      type: object
      additionalProperties: false
      properties:
        user: { $ref: '#/components/schemas/UserSummary' }
        accessToken: { type: string }
        refreshToken: { type: string }
        mfaRequired: { type: boolean }
        mfaTicket: { type: string }
    Library:
      type: object
      additionalProperties: false
      required: [id, name, description, allowReaderUploads, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        allowReaderUploads: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    VerifyPairingResponse:
      type: object
      additionalProperties: false
      required: [enrolmentGrant, address, hostName]
      properties:
        enrolmentGrant: { type: string }
        address: { type: string }
        hostName: { type: string }
...
security:
  - bearerAuth: []